refactor: move `BeaconToSign` to entities module in signer
