+++ mithril-signer/src/entities/mod.rs
@@ -2,6 +2,8 @@
 //!
 //! This module provide domain entities for the services & state machine.
 
+mod beacon_to_sign;
 mod signer_epoch_settings;
 
+pub use beacon_to_sign::*;
 pub use signer_epoch_settings::*;
